Refactor some code and add basic test for preference-service
